Pre-configured.
Most automake users who distribute things via Github check in a
distribution, not a bare source tree.  Intel didn't get the memo.

Do not rebase this commit if ISA-l is updated upstream.  Reset master to the Intel master, then rebase this onto the new master.  Update it as follows:

1. 'autoreconf --force --install'

2. './configure'

3. 'make distdir' (This will create a directory 'isa-l-X.XX.XX' where
X.XX.XX is the version.)

4. 'cd isa-l-X.XX.XX; find -L . -type f > ../files.txt'

5. 'cd ..; rm -rf isa-l-X.XX.XX; xargs git add -f < files.txt; rm files.txt'

6. 'git commit --amend --no-edit'

Signed-off-by: Robert Menke <rob@lightbitslabs.com>